+++ docs/Whats-New.md
@@ -73,6 +73,7 @@ Fixes:
 - Fix alternative war factory selection not respecting the `Naval=` key (by Rampastring)
 - Introduce -4 as a possible argument meaning "any human player" for the "Entered By" event (by Rampastring)
 - Fix crash when loading a game when a game object's or animation's graphics are present in side-specific MIX file (by Rampastring)
+- Add support for loading OwnerDraw (Options Menu) graphics from side-specific MIX files (by Rampastring)
 
 :::